requirements.txt
# Essential dependencies - Compatible with Python 3.12
flask>=2.0.0
anthropic>=0.3.0
requests>=2.28.0
spacy>=3.4.0
numpy>=1.21.0
pandas>=1.3.0
python-docx>=0.8.0
web3>=6.0.0
py-solc-x>=1.12.0
python-dotenv>=0.19.0
textstat>=0.7.0
nltk>=3.7

# PDF Processing
pypdf>=3.0.0

# OCR Dependencies (for scanned PDFs)
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.0.0

# Machine Learning for clause classification
torch>=1.9.0
transformers>=4.20.0
scikit-learn>=1.0.0

# Topic modeling for Balaji contextual alignment metric
gensim>=4.3.0

# Optional - only if you need advanced ML features 